pyproject.toml
[project]
name = "gitlab-flake-bot"
version = "0.1.0"
description = "GitLab bot to update NixOS Flakes"
authors = [{ name = "Dustin Frisch", email = "fooker@lab.sh" }]
readme = "README.md"

requires-python = ">=3.13"
dependencies = [
    "structlog>=25.4.0",
    "pydantic>=2.8.1",
    "pydantic-settings>=2.8.1",
    "python-gitlab>=5.6.0",
    "sh>=2.2.2",
    "munch>=4.0.0",
    "pytimeparse2>=1.7.1",
    "typer>=0.20.0",
]

[project.scripts]
gitlab-flake-bot = "gitlab_flake_bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
dev-dependencies = [
    "pytest",
    "ruff",
    "testcontainers",
]

[tool.ruff]
target-version = "py313"
line-length = 140
